refactor: private view monitor, clearRange guard comment, single-run prefill concurrency test
Address CodeRabbit review nitpicks:
- AbstractIndicator.getBarSeries uses a private readOnlySeriesLock monitor
  instead of synchronized(this), matching the private-monitor pattern used by
  CachedIndicator.
- CachedBuffer.clearRange documents why the trailing break guards the loop
  against i++ overflow at Integer.MAX_VALUE.
- The heavyweight 200k-bar recursive prefill concurrency test moves to a
  dedicated non-parameterized class so it runs once per suite instead of once
  per NumFactory.
