+++ ta4j-core/src/test/java/org/ta4j/core/indicators/CachedIndicatorTest.java
@@ -28,7 +28,6 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicReference;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.*;
@@ -250,63 +249,6 @@ public class CachedIndicatorTest extends AbstractIndicatorTest<Indicator<Num>, N
         assertEquals(countAfterFirstBarRefresh, indicator.getCalculationCount());
     }
 
-    @Test
-    public void recursivePrefillSurvivesConcurrentSeriesRevisionChange() throws Exception {
-        // A recursive indicator's iterative prefill must survive a concurrent series
-        // revision change. The snapshot reconciliation introduced with the moving-
-        // series cache sync truncates the ring buffer from within the nested
-        // getValue() calls of an in-flight prefill; the gap cannot be refilled
-        // iteratively (the prefill depth guard skips the nested prefill), and the
-        // recursive fallback then walks the gap index by index until the stack
-        // overflows.
-        int barCount = 200_000;
-        int latchIndex = 150_000;
-        ConcurrentBarSeries series = new ConcurrentBarSeriesBuilder().withNumFactory(numFactory).build();
-        for (int i = 0; i < barCount; i++) {
-            series.barBuilder()
-                    .timePeriod(Duration.ofMinutes(1))
-                    .endTime(Instant.EPOCH.plus(Duration.ofMinutes(i)))
-                    .closePrice(1)
-                    .add();
-        }
-
-        LatchingSelfReferencingIndicator indicator = new LatchingSelfReferencingIndicator(series, latchIndex);
-        ExecutorService pool = Executors.newFixedThreadPool(2);
-        try {
-            AtomicReference<Throwable> readerFailure = new AtomicReference<>();
-            Future<?> reader = pool.submit(() -> {
-                try {
-                    indicator.getValue(barCount - 1);
-                } catch (Throwable t) {
-                    readerFailure.set(t);
-                }
-            });
-
-            // Wait until the prefill is in flight at the latch index, then mutate a
-            // bar BELOW the prefill position so the next nested getValue() observes a
-            // revision change and the reconciliation would punch a gap under the
-            // in-flight prefill.
-            assertTrue("prefill never reached the latch index", indicator.reached.await(120, TimeUnit.SECONDS));
-            Bar bar = series.getBar(latchIndex);
-            Bar replacement = series.barBuilder()
-                    .timePeriod(bar.getTimePeriod())
-                    .endTime(bar.getEndTime())
-                    .openPrice(500)
-                    .highPrice(500)
-                    .lowPrice(500)
-                    .closePrice(500)
-                    .volume(bar.getVolume())
-                    .build();
-            series.replaceBar(latchIndex - 50_000, replacement);
-
-            reader.get(120, TimeUnit.SECONDS);
-            assertNull("recursive read crashed: " + readerFailure.get(), readerFailure.get());
-            assertNumEquals(barCount, indicator.getValue(barCount - 1));
-        } finally {
-            pool.shutdownNow();
-        }
-    }
-
     private static final class FirstBarReadingIndicator extends CachedIndicator<Num> {
 
         private final AtomicInteger calculationCount = new AtomicInteger();
@@ -323,33 +265,6 @@ public class CachedIndicatorTest extends AbstractIndicatorTest<Indicator<Num>, N
 
         int getCalculationCount() {
             return calculationCount.get();
-        }
-
-        @Override
-        public int getCountOfUnstableBars() {
-            return 0;
-        }
-    }
-
-    private static final class LatchingSelfReferencingIndicator extends RecursiveCachedIndicator<Num> {
-
-        private final int latchIndex;
-        private final CountDownLatch reached = new CountDownLatch(1);
-
-        private LatchingSelfReferencingIndicator(BarSeries series, int latchIndex) {
-            super(series);
-            this.latchIndex = latchIndex;
-        }
-
-        @Override
-        protected Num calculate(int index) {
-            if (index == latchIndex) {
-                reached.countDown();
-            }
-            if (index == 0) {
-                return getBarSeries().numFactory().one();
-            }
-            return getValue(index - 1).plus(getBarSeries().numFactory().one());
         }
 
         @Override
